feat: Update location and export of HAM model
