+++ lib/src/cmaf-ham.ts
@@ -3,8 +3,8 @@
  *
  * @packageDocumentation
  */
-export {Presentation} from './cmaf/ham/Presentation';
-export {SwitchingSet} from './cmaf/ham/SwitchingSet';
-export {Track} from './cmaf/ham/Track';
-export {m3u8toHam} from './cmaf/ham/manifestParser';
-export {parseM3u8} from './cmaf/utils/m3u8';
+export { Presentation } from './cmaf/ham/model/Presentation';
+export { SwitchingSet } from './cmaf/ham/model/SwitchingSet';
+export { Track } from './cmaf/ham/model/Track';
+export { m3u8toHam } from './cmaf/ham/manifestParser';
+export { parseM3u8 } from './cmaf/utils/m3u8';
