--- lib/src/cmaf/ham/SelectionSet.ts
@@ -1,13 +1,0 @@
-import { SwitchingSet } from './SwitchingSet';
-
-export class SelectionSet {
-	id: string;
-	duration: number;
-	switchingSet: SwitchingSet[];
-
-	constructor(id: string, duration: number, switchingSet: SwitchingSet[]) {
-		this.id = id;
-		this.duration = duration;
-		this.switchingSet = switchingSet;
-	}
-}
