--- lib/src/cmaf/ham/SwitchingSet.ts
@@ -1,20 +1,0 @@
-import { Track } from './Track';
-
-export class SwitchingSet {
-
-	id: string;
-	type: string;
-	codec: string;
-	duration: number;
-	language: string;
-	tracks: Track[];
-
-	constructor(id: string, type: string, codec: string, duration: number, language: string, tracks: Track[]) {
-		this.id = id;
-		this.type = type;
-		this.codec = codec;
-		this.duration = duration;
-		this.language = language;
-		this.tracks = tracks;
-	}
-}
