+++ lib/src/cmaf/ham/model/Segment.ts
